separate utilities functions to themes
