+++ wombat/utilities/utilities.py
@@ -1,18 +1,12 @@
-"""Provides various utility functions."""
+"""Provides various utility functions that don't fit within a common theme."""
 
 
 from __future__ import annotations
 
-import logging  # type: ignore
-import datetime
-from typing import Any, Callable  # type: ignore
-from pathlib import Path  # type: ignore
-from logging.handlers import MemoryHandler
+from typing import Callable  # type: ignore
 
 import numpy as np  # type: ignore
 import pandas as pd
-
-from wombat.core.data_classes import HOURS_IN_DAY  # type: ignore
 
 
 try:
@@ -21,113 +15,6 @@ except ImportError:
     from functools import lru_cache
 
     cache = lru_cache(None)
-
-
-def convert_dt_to_hours(diff: datetime.timedelta) -> float:
-    """Converts a ``datetime.timedelta`` object to number of hours at the seconds resolution.
-
-    Parameters
-    ----------
-    diff : datetime.timedelta
-        The difference between two ``datetime.datetime`` objects.
-
-    Returns
-    -------
-    float
-        Number of hours between to ``datetime.datetime`` objects.
-    """
-    days = diff.days * 24 if diff.days > 0 else 0
-    seconds = diff.seconds / 60 / 60
-    return days + seconds
-
-
-def hours_until_future_hour(dt: datetime.datetime, hour: int) -> float:
-    """Number of hours until a future hour in the same day for ``hour`` <= 24, otherwise,
-    it is the number of hours until a time in the proceeding days.
-
-    Parameters
-    ----------
-    dt : datetime.datetime
-        Focal datetime.
-    hour : int
-        Hour that is later in the day, in 24 hour time.
-
-    Returns
-    -------
-    float
-        Number of hours between the two times.
-    """
-    if hour >= 24:
-        days, hour = divmod(hour, 24)
-
-        # Convert to python int
-        days = int(days)
-        hour = int(hour)
-
-        new_dt = dt + datetime.timedelta(days=days)
-        new_dt = new_dt.replace(hour=hour, minute=0, second=0)
-    else:
-        new_dt = dt.replace(hour=hour, minute=0, second=0)
-    diff = new_dt - dt
-    return convert_dt_to_hours(diff)
-
-
-def check_working_hours(
-    env_start: int, env_end: int, workday_start: int, workday_end: int
-) -> tuple[int, int]:
-    """Checks the working hours of a port or servicing equipment, and overrides a
-    default (-1) to the environment's settings, otherwise returns back the input hours.
-
-    Parameters
-    ----------
-    env_start : int
-        The starting hour for the environment's shift
-    env_end : int
-        The ending hour for the environment's shift
-    workday_start : int
-        The starting hour to be checked.
-    workday_end : int
-        The ending hour to be checked.
-
-    Returns
-    -------
-    tuple[int, int]
-        The starting and ending hour to be applied back to the port or servicing equipment.
-    """
-    start_is_invalid = workday_start == -1
-    end_is_invalid = workday_end == -1
-
-    start = env_start if start_is_invalid else workday_start
-    end = env_end if end_is_invalid else workday_end
-
-    return start, end
-
-
-def calculate_cost(
-    duration: int | float, rate: float, n_rate: int = 1, daily_rate: bool = False
-) -> float:
-    """Calculates the equipment cost, or labor cost for either salaried or hourly employees.
-
-    Parameters
-    ----------
-    duration : int | float
-        Length of time, in hours.
-    rate : float
-        The labor or equipment rate, in $USD/hour.
-    n_rate : int
-        Total number of of the rate to be applied, more than one if``rate`` is broken
-        down by number of individual laborers (if rate is a labor rate), by default 1.
-    daily_rate : bool, optional
-        Indicator for if the ``rate`` is a daily rate (True), or hourly rate (False), by
-        default False.
-
-    Returns
-    -------
-    float
-        The total cost of the labor performed.
-    """
-    multiplier = duration / HOURS_IN_DAY if daily_rate else duration
-    return multiplier * rate * n_rate
 
 
 @cache
@@ -145,133 +32,6 @@ def _mean(*args) -> float:
         The average of the values provided
     """
     return np.mean(args)
-
-
-def setup_logger(
-    logger_name: str, log_file: Path, level: Any = logging.INFO, capacity: int = 500
-) -> None:
-    """Creates the logging infrastructure for a given logging category.
-
-    TODO: Figure out how to type check ``logging.INFO``; ``Callable``?
-
-    Parameters
-    ----------
-    logger_name : str
-        Name to assign to the logger.
-    log_file : Path
-        File name and path for where the log data should be saved.
-    level : Any, optional
-        Logging level, by default logging.INFO.
-    """
-    logger = logging.getLogger(logger_name)
-    formatter = logging.Formatter(
-        "%(asctime)s :: %(name)s :: %(levelname)s :: %(message)s"
-    )
-    fileHandler = logging.FileHandler(log_file, mode="w")
-    fileHandler.setFormatter(formatter)
-
-    memory_handler = MemoryHandler(capacity=capacity, target=fileHandler)
-    memory_handler.setFormatter(formatter)
-
-    logger.setLevel(level)
-    # logger.addHandler(fileHandler)
-    logger.addHandler(memory_handler)
-
-
-def format_events_log_message(
-    simulation_time: datetime.datetime,
-    env_time: float,
-    system_id: str,
-    system_name: str,
-    part_id: str,
-    part_name: str,
-    system_ol: float | str,
-    part_ol: float | str,
-    agent: str,
-    action: str,
-    reason: str,
-    additional: str,
-    duration: float,
-    request_id: str,
-    materials_cost: int | float = 0,
-    hourly_labor_cost: int | float = 0,
-    salary_labor_cost: int | float = 0,
-    equipment_cost: int | float = 0,
-) -> str:
-    """Formats the logging messages into the expected format for logging.
-
-    Parameters
-    ----------
-    simulation_time : datetime64
-        Timestamp within the simulation time.
-    env_time : float
-        Environment simulation time (``Environment.now``).
-    system_id : str
-        Turbine ID, ``System.id``.
-    system_name : str
-        Turbine name, ``System.name``.
-    part_id : str
-        Subassembly, component, or cable ID, ``_.id``.
-    part_name : str
-        Subassembly, component, or cable name, ``_.name``.
-    system_ol : int | float
-        System operating level, ``System.operating_level``. Use an empty string for n/a.
-    part_ol : int | float
-        Subassembly, component, or cable operating level, ``_.operating_level``. Use an
-        empty string for n/a.
-    agent : str
-        Agent performin the action.
-    action : str
-        Action that was taken.
-    reason : str
-        Reason an action was taken.
-    additional : str
-        Any additional information that needs to be logged.
-    duration : float
-        Length of time the action lasted.
-    request_id : str
-        The ``RepairRequest.request_id`` or "na".
-    materials_cost : int | float, optional
-        Total cost of materials for action, in USD, by default 0.
-    hourly_labor_cost : int | float, optional
-        Total cost of hourly labor for action, in USD, by default 0.
-    salary_labor_cost : int | float, optional
-        Total cost of salaried labor for action, in USD, by default 0.
-    equipment_cost : int | float, optional
-        Total cost of equipment for action, in USD, by default 0.
-
-    Returns
-    -------
-    str
-        Formatted message for consistent logging.[summary]
-    """
-    total_labor_cost = hourly_labor_cost + salary_labor_cost
-    total_cost = total_labor_cost + equipment_cost + materials_cost
-    message = " :: ".join(
-        (
-            f"{simulation_time}",
-            f"{env_time:f}",
-            f"{system_id}",
-            f"{system_name}",
-            f"{part_id}",
-            f"{part_name}",
-            f"{system_ol:f}",
-            f"{part_ol:f}",
-            f"{agent}",
-            f"{action}",
-            f"{reason}",
-            f"{additional}",
-            f"{duration:f}",
-            f"{request_id}",
-            f"{materials_cost:f}",
-            f"{hourly_labor_cost:f}",
-            f"{salary_labor_cost:f}",
-            f"{equipment_cost:f}",
-            f"{total_labor_cost:f}",
-            f"{total_cost:f}",
-        )
-    )
-    return message
 
 
 def IEC_power_curve(
